pyproject.toml
[project]
name = "voicebridge"
version = "0.1.0"
description = "Telegram bot that translates voice messages across Uzbek, Russian, and English."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiogram>=3.18.0,<4.0.0",
  "faster-whisper>=1.1.1,<2.0.0",
  "gTTS>=2.5.4,<3.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "sentencepiece>=0.2.0,<1.0.0",
  "transformers>=4.57.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.25.0,<1.0.0",
  "ruff>=0.11.0,<0.12.0",
]
uzbek_asr = [
  "nemo_toolkit[asr]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voicebridge"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
]
